Refactor phase 4a: split system.py into the app_utils/system package
`app_utils/system.py` was 2,580 lines — the Phase 4 entry the refactor
plan flagged as "mostly independent helpers — easier than it looks".
It was: 47 top-level definitions sharing one file, a strictly acyclic
internal call graph, and no `__file__`-relative paths to shift.

It is now `app_utils/system/`, 16 modules behind a re-exporting
`__init__`: common, dependencies, services, badges, osinfo, network,
device_tree, block_devices, hardware, disks, smart, smart_fields,
temperature, rtc, subsystems, snapshot. 14 of the 16 are under the
400-line guidance.

Pure motion: 48 top-level definitions, 48 `ast.dump()` matches, zero
differences, plus the standing assertion that every non-blank line of
the original lands in exactly one module. All four production consumers
import cleanly (`app_utils/__init__.py`, `app_core/system_health.py`,
`webapp/admin/api.py`, `scripts/diagnose_smart.sh`).

`DEVICE_TREE_CANDIDATES` is deliberately not re-exported from the
package — it is a mutable list two tests replace with
`monkeypatch.setattr`, and re-exporting it would have made both patches
silent no-ops. Leaving it out failed loudly, which is what pointed at
the right target; the same applies to `Path` and `psutil` in the
CPU-details test. All three now patch `app_utils.system.<module>`.

`docs/guides/SMART_SETUP.md` pointed at line ranges in the old monolith
and at `webapp/routes_public.py` for the health-page route, which moved
in 2.141.0; both now name modules and functions.

`snapshot.py` (478) and `smart.py` (429) stay over the cap because each
is one function — 406 and 396 lines. They need collaborator extraction
with a characterization harness, tracked as Phase 4a-ii.

Tests: tests/test_system_health_utils.py,
tests/test_system_health_fixes.py,
tests/test_system_health_template_fields.py (43 passed); full suite
2051 passed, 67 xfailed.
